added layout, moved all styling to /css + some styling changes
