+++ src/ItemList.jsx
@@ -1,5 +1,5 @@
 import { useState, useMemo } from 'react';
-import './ItemList.css';
+import './css/ItemList.css';
 
 // sample world wonders; real data will come from the API/database later
 const initialItems = [
