+++ src/css/layout.css
@@ -1,0 +1,63 @@
+/* GENERAL PAGE LAYOUT */
+/* .layout {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+
+} */
+
+
+/* NAVBAR SECTION */
+.navbar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+
+  padding: 16px;
+  background-color: #000000;
+  color: white;
+}
+
+
+/* SITE TITLE */
+.logo {
+  margin: 0;
+}
+
+
+/* NAV BUTTON AREA */
+.nav-links {
+  display: flex;
+  gap: 10px;
+}
+
+
+/* NAV BUTTON STYLE */
+.nav-links button {
+  padding: 6px 12px;
+  border: none;
+  background-color: #555;
+  color: white;
+  cursor: pointer;
+}
+
+.nav-links button:hover {
+  background-color: #777;
+}
+
+
+/* MAIN CONTENT AREA */
+.content {
+  flex: 1; /* Take up remaining space */  
+  padding: 20px;
+  max-width: 1200px;
+  margin: 0 auto;
+}
+
+
+/* FOOTER SECTION */
+.footer {
+  text-align: center;
+  padding: 10px;
+  background-color: #000000;
+}
